Update to java 17
As this artifact is used by the reports this needs to be updated
to java 17.
As per the current plan reports built with java 17 will be tested
in canary so there will be a period where both old reports built
with java 8 and new reports built with java 17 needs to be available
so sstemplates built with java 8 is also needs to be available.
So updating the version of the artifacts to 2.0.0

1.0.0-SNAPSHOTS are built with java 8 and 2.0.0-SNAPSHOTS will
be built with java 17