Update to javax.el
Apache commons-el is end of line and using depricated EL (jsp-api)
This commit is replacing commons-el to juel library which
supports javax.el, so that we can upgrade java, spring and other
to the latest.

One notable change
commons-el was resolving ${param.x} to null even if param bean
was not initialized but the juel was throwing nosuchproperty
exception and sstemplates was using param so in the variable
mapper param is added as a variable with empty map while it
is initialized. There won't be any side effects due to this
and this is done so that we don't need to change any core
logic in sstemplates.